Fix: produits producteur + verifs et erreurs

+++ source/Controller/ProducteurController.php
@@ -31,7 +31,7 @@ class ProducteurController extends UserController
             //* Valide les données du formulaire
             $datas = $this->validate($_POST, [
                 'DesignationProduitProducteur', 'PrixProduitProducteur',
-                'DetailsProduitProducteur', 'QuantiteProduitProducteur', 'IdProduitProduitProducteur', 'ImageProduitProducteur'
+                'DetailsProduitProducteur', 'QuantiteProduitProducteur', 'IdProduitProduitProducteur'
             ]);
             $idProducteur = $_SESSION['user']['IdRole'];
 
@@ -57,34 +57,47 @@ class ProducteurController extends UserController
                 } else {
                     $upload_path = "assets/images/fruit.jpg";
                 }
-                //* Crée un modèle ProduitProducteur
-                $ProduitProducteur = new ProduitProducteurModel();
-                //* Remplit les propriétés du modèle avec les données du formulaire
-                $ProduitProducteur->IdProducteurProduitProducteur = $idProducteur;
-                $ProduitProducteur->IdProduitProduitProducteur  = $datas['IdProduitProduitProducteur'];
-                $ProduitProducteur->DesignationProduitProducteur = htmlentities($datas['DesignationProduitProducteur'], ENT_QUOTES);
-                $ProduitProducteur->PrixProduitProducteur = $datas['PrixProduitProducteur'];
-                $ProduitProducteur->DetailsProduitProducteur = $datas['DetailsProduitProducteur'];
-                $ProduitProducteur->QuantiteProduitProducteur = $datas['QuantiteProduitProducteur'];
-                $ProduitProducteur->ImageProduitProducteur = $upload_path;
 
-                //* Enregistre les données dans la base de données
-                $IdProduitProducteur = $ProduitProducteur->Save();
+                if ($datas['QuantiteProduitProducteur'] < 10) {
+                    ExceptionHandler::SetUserError('Quantité trop faible pour la mise en vente');
+                }
 
-                $Demandes = new DemandesModel();
-                $Demandes->IdUserDemande = $_SESSION['user']['IdRole'];
-                $Demandes->IdProduitProducteurDemande = $IdProduitProducteur;
-                $Demandes->ObjetDemande = 'Ajout';
-                $Demandes->PrixProposeDemande = $datas['PrixProduitProducteur'];
-                $Demandes->DesignationProduitDemande = htmlentities($datas['DesignationProduitProducteur'], ENT_QUOTES);
-                $Demandes->MotifDemande = "Le producteur " . $_SESSION['user']['Username'] . " souhaite ajouter le produit " . htmlentities($datas['DesignationProduitProducteur'], ENT_QUOTES) .
-                    " au prix de " . $datas['PrixProduitProducteur'];
-                $Demandes->Save();
+                if ($datas['PrixProduitProducteur'] <= 0) {
+                    ExceptionHandler::SetUserError('Veuillez entrer un prix réel');
+                }
+
+                $errors = ExceptionHandler::GetUserError();
+
+                if (count($errors) == 0) {
+                    //* Crée un modèle ProduitProducteur
+                    $ProduitProducteur = new ProduitProducteurModel();
+                    //* Remplit les propriétés du modèle avec les données du formulaire
+                    $ProduitProducteur->IdProducteurProduitProducteur = $idProducteur;
+                    $ProduitProducteur->IdProduitProduitProducteur  = $datas['IdProduitProduitProducteur'];
+                    $ProduitProducteur->DesignationProduitProducteur = htmlentities($datas['DesignationProduitProducteur'], ENT_QUOTES);
+                    $ProduitProducteur->PrixProduitProducteur = $datas['PrixProduitProducteur'];
+                    $ProduitProducteur->DetailsProduitProducteur = $datas['DetailsProduitProducteur'];
+                    $ProduitProducteur->QuantiteProduitProducteur = $datas['QuantiteProduitProducteur'];
+                    $ProduitProducteur->ImageProduitProducteur = $upload_path;
+
+                    //* Enregistre les données dans la base de données
+                    $IdProduitProducteur = $ProduitProducteur->Save();
+
+                    $Demandes = new DemandesModel();
+                    $Demandes->IdUserDemande = $_SESSION['user']['IdRole'];
+                    $Demandes->IdProduitProducteurDemande = $IdProduitProducteur;
+                    $Demandes->ObjetDemande = 'Ajout';
+                    $Demandes->PrixProposeDemande = $datas['PrixProduitProducteur'];
+                    $Demandes->DesignationProduitDemande = htmlentities($datas['DesignationProduitProducteur'], ENT_QUOTES);
+                    $Demandes->MotifDemande = "Le producteur " . $_SESSION['user']['Username'] . " souhaite ajouter le produit " . htmlentities($datas['DesignationProduitProducteur'], ENT_QUOTES) .
+                        " au prix de " . $datas['PrixProduitProducteur'];
+                    $Demandes->Save();
 
 
-                //* Redirige l'utilisateur avec un message de succès
-                header('location: /User/AddProduct?info=Produit ajouté avec succès');
-                exit();
+                    //* Redirige l'utilisateur avec un message de succès
+                    header('location: /User/AddProduct?info=Produit ajouté avec succès');
+                    exit();
+                }
             } else {
                 //* Gère les erreurs de validation
                 ExceptionHandler::SetUserError("Erreur");
@@ -113,8 +126,10 @@ class ProducteurController extends UserController
     //* fonction pour listé la liste des produits du producteurs, qui s'adapte au formulaire de suppresion, et d'update.
     public function ProductList(): void
     {
+        $this->connectCheck('user', 'Producteur', '/User');
 
         $AllBundles = $this->listBundle();
+        $errors = [];
 
         $ProduitsBundle = [];
         if ($AllBundles) {
@@ -126,12 +141,22 @@ class ProducteurController extends UserController
         $this->connectCheck('user', 'Producteur');
         //* Vérifie si l'utilisateur a soumis un formulaire de suppression
         if (isset($_POST['delete'])) {
-           $this->DeleteProductProducteur();
+            if (isset($_POST['deleteProduit'])) {
+                $this->DeleteProductProducteur();
+            } else {
+                ExceptionHandler::SetUserError('Veuillez selectionner au moins 1 produit');
+                $errors = ExceptionHandler::GetUserError();
+            }
         }
 
         //* Vérifie si l'utilisateur a soumis un formulaire de mise à jour
         if (isset($_POST['update'])) {
-            $this->UpdateProductProducteur();
+            if (isset($_POST['MajPrix'])) {
+                $errors = $this->UpdateProductProducteur($_POST);
+            } else {
+                ExceptionHandler::SetUserError('Veuillez selectionner au moins 1 produit');
+                $errors = ExceptionHandler::GetUserError();
+            }
         }
 
         //* Crée une instance de ProduitProducteurModel pour gérer les produits du producteur
@@ -139,6 +164,7 @@ class ProducteurController extends UserController
         $AllProduits = $Produits->getProduitProducteur($_SESSION['user']['IdRole'], true);
 
         ViewController::Set('title', 'Mes produits');
+        ViewController::Set('errors', $errors);
         ViewController::Set('SessionInfo', $_SESSION['user']);
         ViewController::Set('AllBundles', $AllBundles);
         ViewController::Set('ProduitsBundle', $ProduitsBundle);
@@ -149,32 +175,59 @@ class ProducteurController extends UserController
     //* function pour mettre a jour les produits du producteurs, uniquement le prix, elle gére également la demande de modification pour emettre
     //* a l'admin
 
-    private function UpdateProductProducteur(): void
+    private function UpdateProductProducteur($post): bool|array
     {
         //* Récupère l'ID du producteur à partir de la session
         $idProducteur = $_SESSION['user']['IdRole'];
 
         //* Vérifie si un tableau de produits à mettre à jour a été soumis
-        if (isset($_POST['update']) && is_array($_POST['update'])) {
-            foreach ($_POST['update'] as $IdProduitProducteur => $datas) {
+        $datas = $this->validate($post, ['MajPrix', 'Prix']);
+        if ($datas !== false) {
+            for ($i = 0; $i < count($_POST['MajPrix']); $i++) {
                 //* Valide les données du formulaire
-                $datas = $this->validate($datas, ['DesignationProduitProducteur', 'PrixProduitProducteur', 'QuantiteProduitProducteur', 'DetailsProduitProducteur']);
-                if ($datas !== false) {
-                    //* Crée un modèle ProduitProducteur et effectue la mise à jour
+                //* Crée un modèle ProduitProducteur et effectue la mise à jour
+
+                if ($datas['Prix'][$i] <= 0) {
+                    ExceptionHandler::SetUserError('Veuillez entrer un prix réel');
+                }
+
+                $errors = ExceptionHandler::GetUserError();
+
+                if (count($errors) == 0) {
                     $ProduitProducteurModel = new ProduitProducteurModel();
-                    $ProduitProducteurModel->producteurProduitUpdate($datas, $idProducteur, $IdProduitProducteur);
+                    $ProduitProducteurModel->IdProduitProducteur = $datas['MajPrix'][$i];
+                    $ProduitProducteur = $ProduitProducteurModel->Find('*', 'Fetch');
+
+                    if ($ProduitProducteur['PrixProduitProducteur'] == $datas['Prix'][$i]) {
+                        ExceptionHandler::SetUserError('Veuillez entrer un prix différent de l\'ancien');
+                        $errors = ExceptionHandler::GetUserError();
+                        return $errors;
+                    }
+
+                    //* On retire l'ID afin de ne pas l'inclure dans la requete;
+                    unset($ProduitProducteurModel->IdProduitProducteur);
+
+                    $ProduitProducteurModel->PrixProduitProducteur = $datas['Prix'][$i];
+                    $ProduitProducteurModel->IsValidateProduitProducteur = 0;
+                    $ProduitProducteurModel->Where($datas['MajPrix'][$i]);
+                    $ProduitProducteurModel->Update();
 
                     $demandes = new DemandesModel();
                     $demandes->IdUserDemande = $_SESSION['user']['IdRole'];
-                    $demandes->IdProduitProducteurDemande = $IdProduitProducteur;
+                    $demandes->IdProduitProducteurDemande = $_POST['MajPrix'][$i];
                     $demandes->ObjetDemande = 'Prix';
-                    $demandes->DesignationProduitDemande = htmlentities($datas['DesignationProduitProducteur'], ENT_QUOTES);
-                    $demandes->PrixProposeDemande = $datas['PrixProduitProducteur'];
-                    $demandes->MotifDemande = "Le producteur " . $_SESSION['user']['Username'] . " souhaite modifié le prix de " . htmlentities($datas['DesignationProduitProducteur'], ENT_QUOTES) .
-                        " au prix de " . $datas['PrixProduitProducteur'];
+                    $demandes->DesignationProduitDemande = htmlentities($ProduitProducteur['DesignationProduitProducteur'], ENT_QUOTES);
+                    $demandes->PrixProposeDemande = $datas['Prix'][$i];
+                    $demandes->MotifDemande = "Le producteur " . $_SESSION['user']['Username'] . " souhaite modifié le prix de " . htmlentities($ProduitProducteur['DesignationProduitProducteur'], ENT_QUOTES) .
+                        " au prix de " . $datas['Prix'][$i];
                     $demandes->Save();
+                } else {
+                    return $errors;
                 }
             }
+            header('Location: /User/ProductList');
+            exit();
+            return true;
         }
     }
 
@@ -182,29 +235,56 @@ class ProducteurController extends UserController
 
     private function DeleteProductProducteur(): void
     {
-        // FIX VOIR AVEC NICOLAS POUR LES FORREIGN KEY
         // * Vérifie si un formulaire de suppression a été soumis
-if (isset($_POST['delete'])) {
-  
-        if (isset($_POST['deleteProduit']) && is_array($_POST['deleteProduit'])) {
-            $Produit = new ProduitProducteurModel();
-             // Parcourt les produits à supprimer et les supprime un par un
-            foreach ($_POST['deleteProduit'] as $IdProduitProducteur) {
-                
-                //* Supprime le produit correspondant
-                $ProduitInfo = $Produit->getOneProduitProducteur($IdProduitProducteur);
-                $imagePath = $ProduitInfo['ImageProduitProducteur'];
 
-                //* Supprime le fichier image associé
-                if (file_exists($imagePath)) {
-                    unlink(DIR_PUBLIC . $imagePath);
-                }
-                //* Supprime le produit correspondant
-                $ProduitInfo->Delete();
+        $Produit = new ProduitProducteurModel();
+        // Parcourt les produits à supprimer et les supprime un par un
+        foreach ($_POST['deleteProduit'] as $IdProduitProducteur) {
+
+            // ? Pour le futur, créer un controller chargé de gérer les suppressions en fonction des contraintes de FK plutôt que de recopier;
+            // NOTICE Voir admin pour commentaires
+
+            $demandes = new DemandesModel();
+            $demandes->IdProduitProducteurDemande = $IdProduitProducteur;
+
+            $result = $demandes->Find('IdDemande');
+            if ($result) {
+                $demandes->Delete();
             }
-        } 
-    } else {
-            echo "une erreur est survenue";
+
+            $Bundle = new BundleModel();
+            $IdProduitsBundle = $Bundle->Find("IdProduitsBundle, IdBundle");
+            $IdProduitsBundleArr = [];
+
+            foreach ($IdProduitsBundle as $ProduitsBundle) {
+                array_push($IdProduitsBundleArr, explode(',', $ProduitsBundle['IdProduitsBundle']));
+            }
+
+            foreach ($IdProduitsBundleArr as $key => $Arrays) {
+                foreach ($Arrays as $Id) {
+                    if ($Id === $IdProduitProducteur) {
+                        $Bundle->IdBundle = $IdProduitsBundle[$key]['IdBundle'];
+                        $Bundle->Delete();
+                        break;
+                    }
+                }
+            }
+
+            //* Supprime le produit correspondant
+            $Produit->IdProduitProducteur = $IdProduitProducteur;
+            $ProduitInfo = $Produit->Find('ImageProduitProducteur', 'Fetch');
+            $imagePath = $ProduitInfo['ImageProduitProducteur'];
+
+            //* Supprime le fichier image associé
+            if (file_exists($imagePath)) {
+                unlink(DIR_PUBLIC . $imagePath);
+            }
+            //* Supprime le produit correspondant
+
+            $Produit->Delete();
+
+            header('Location: /User/ProductList');
+            exit();
         }
     }
